+++ tedi/components/helpers/index.ts
@@ -1,4 +1,5 @@
 export * from "./empty-state";
+export * from "./ellipsis";
 export * from "./grid";
 export * from "./scroll-fade";
 export * from "./separator/separator.component";